Move protocol logic to new `protocol` module
The `vdaf` module currently implements the core protocol logic for DAP,
much of which has little to do with VDAFs directly. To make the code
easier to maintain, move this code to its own module called `protocol`.

This module has three submodules, one for each of the DAP roles:
`client`, `aggregator`, and `collector`.

The new module is crate-private. This change to the API does not impact
the other crates in this workspace, with the exception of some report
state internals needed by `daphne_server`: these are `EarlyReportState`,
`EarlyReportStateConsumed`, and `EarlyReportStateInitialized`. We use
`pub use` to export these.

There are no functional changes in this commit; it is limited to code
movement only.
