+++ daphne/src/roles/helper.rs
@@ -19,8 +19,8 @@ use crate::{
         ReportId, TaskId, TransitionFailure, TransitionVar,
     },
     metrics::{DaphneMetrics, DaphneRequestType},
+    protocol::aggregator::ReportProcessedStatus,
     roles::aggregator::MergeAggShareError,
-    vdaf::ReportProcessedStatus,
     DapAggregateShare, DapAggregateSpan, DapAggregationJobState, DapError,
     DapHelperAggregationJobTransition, DapRequest, DapResource, DapResponse, DapTaskConfig,
     DapVersion, MetaAggregationJobId,
